получение offers и destinations с сервера
