--- src/destination-api-service.js
@@ -1,8 +1,0 @@
-import ApiService from './framework/api-service.js';
-
-export default class DestinationApiService extends ApiService {
-  get destinations() {
-    return this._load({ url: 'destinations' })
-      .then(ApiService.parseResponse);
-  }
-}
