+++ src/point-api-service.js
@@ -11,6 +11,16 @@ export default class PointApiService extends ApiService {
       .then(ApiService.parseResponse);
   }
 
+  get offers() {
+    return this._load({ url: 'offers' })
+      .then(ApiService.parseResponse);
+  }
+
+  get destinations() {
+    return this._load({ url: 'destinations' })
+      .then(ApiService.parseResponse);
+  }
+
   updatePoint = async (point) => {
     const response = await this._load({
       url: `points/${point.id}`,
